Learned new stuff

+++ Fast-API2/myapi.py
@@ -1,14 +1,29 @@
-from fastapi import FastAPI, HTTPException, Depends
-from sqlalchemy import create_engine, Column, Integer, String
-from sqlalchemy.ext.declarative import declarative_base
+from fastapi import FastAPI, HTTPException, Depends, status
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from sqlalchemy import create_engine, Column, Integer, String, Boolean
+from sqlalchemy.orm import declarative_base
 
 from sqlalchemy.orm import sessionmaker, Session
 
 from pydantic import BaseModel
 from typing import Optional, List
 
+from passlib.context import CryptContext
+import jwt
+from jwt.exceptions import InvalidTokenError
 
-app = FastAPI(title="Integration with SQL - New Project")
+from datetime import datetime,timedelta
+
+
+# Security config
+SECRET_KEY = "aryan"
+ALGORITHM = "HS256"
+TOKEN_EXPIRES = 30
+
+#password context
+pwd_context = CryptContext(schemes=['bcrypt'], deprecated = "auto")
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
 
 #Database Setup
 engine = create_engine("sqlite:///users.db", connect_args={"check_same_thread":False})
@@ -23,6 +38,8 @@ class User(Base):
     name= Column(String(100), nullable=False)
     email = Column(String(50), nullable=False, unique=True)
     role = Column(String(50), nullable=False)
+    hashed_pwd = Column(String, nullable=False)
+    is_active = Column(Boolean, default=True)
 
 Base.metadata.create_all(engine)
 
@@ -31,15 +48,70 @@ class UserCreate(BaseModel):
     name:str
     email:str
     role:str
+    password:str
+
 
 class UserResponse(BaseModel):
     id:int
     name:str
     email: str
     role: str
+    is_active:bool
 
     class Config:
         from_attributes = True
+
+# New Pydantic Model
+class UserLogin(BaseModel):
+    email:str
+    password:str
+
+class Token(BaseModel):
+    access_token:str
+    token_type:str
+
+class TokenData(BaseModel):
+    email: Optional[str]=None
+
+#Security Functions
+def verify_pwd(plain_pwd:str, hashed_pwd:str)->bool:
+    return pwd_context.verify(plain_pwd, hashed_pwd)
+
+def get_pwd_hash(password:str)->str:
+    return pwd_context.hash(password)
+
+def create_access_token(data:dict,expires_delta:Optional[timedelta]=None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.utcnow()+ expires_delta
+
+    else:
+        expire = datetime.utcnow()+ timedelta(minutes=15)
+    
+    to_encode.update({"exp":expire})
+    encode_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+
+    return encode_jwt
+
+
+def verify_token(token:str)->TokenData:
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        email:str = payload.get("sub")
+        if email is None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail = "Could not verify credentials",
+                headers = {"WWW-Authenticate": "Bearer"}
+            )
+        return TokenData(email=email)
+    except InvalidTokenError:
+        raise HTTPException(
+            status_code=401,
+            detail="Could not verify credentials",
+            headers={"WWW-Authenticate": "Bearer"}
+    )
+
 
 def get_db():
     db = sessionLocal()
@@ -47,41 +119,133 @@ def get_db():
         yield db
     finally:
         db.close()
-get_db()
+
+#auth Dependency
+def get_current_user(token:str= Depends(oauth2_scheme), db: Session = Depends(get_db)):
+    token_data = verify_token(token)
+    user = db.query(User).filter(User.email==token_data.email).first()
+    if user is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail = "User dosn't exist!",
+            headers = {"WWW-Authenticate": "Bearer"}
+        )
+    return user
+
+def get_current_active_user(current_user:User=Depends(get_current_user)):
+    if not current_user.is_active:
+        raise HTTPException(
+            status_code=403,
+            detail = "Inactive User",
+        )
+    return current_user
+
+
+app = FastAPI(title="Integration with SQL - New Project")
+
+#Auth EndPoints
+@app.post("/register", response_model=UserResponse)
+def register_user(user: UserCreate, db: Session=Depends(get_db)):
+    if db.query(User).filter(User.email==user.email).first():
+        raise HTTPException(
+            status_code=409,
+            detail = "User Already Created!",
+        )
+    
+    hashed_password = get_pwd_hash(user.password)
+    db_user = User(
+        name = user.name,
+        email = user.email,
+        role = user.role,
+        hashed_pwd= hashed_password
+    )
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
+
+@app.post("/token", response_model=Token)
+def login_for_access_token(form_data: OAuth2PasswordRequestForm=Depends(), db:Session=Depends(get_db)):
+    user = db.query(User).filter(User.email==form_data.username).first()
+
+    if not user or not verify_pwd(form_data.password, user.hashed_pwd):
+        raise HTTPException(
+            status_code=401,
+            detail = "Wrong email or password",
+        )
+
+    if not user.is_active:
+        raise HTTPException(
+            status_code=403,
+            detail = "Inactive User",
+        )
+
+    access_token_expires = timedelta(minutes=TOKEN_EXPIRES)
+    access_token = create_access_token(
+        data={"sub":user.email}, expires_delta=access_token_expires
+    )
+    return {"access_token":access_token, "token_type":"bearer"}
 
 # Endpoints (www.example.com/)
 @app.get("/")
 def root():
     return {"message": "Intro to FastAPI"}
 
+@app.get("/profile", response_model=UserResponse)
+def get_profile(current_user:User=Depends(get_current_active_user)):
+    return current_user
+
+@app.get("/verify-token")
+def verify_token_endpoints(current_user:User=Depends(get_current_active_user)):
+    return {
+        "valid":True,
+        "user": {
+            "id":current_user.id,
+            "name":current_user.name,
+            "email":current_user.email,
+            "role": current_user.role
+        }
+    }
+
+
+
 @app.get("/users/{user_id}",response_model=UserResponse)
-def get_user(user_id:int, db:Session=Depends(get_db)):
+def get_user(user_id:int,current_user:User=Depends(get_current_active_user),db:Session=Depends(get_db)):
     user = db.query(User).filter(User.id==user_id).first()
     if not user:
         raise HTTPException(status_code=404, detail="User not found!")
     return user
 
 @app.post("/users/", response_model=UserResponse)
-def create_user(user:UserCreate, db:Session=Depends(get_db)):
+def create_user(user:UserCreate,current_user:User=Depends(get_current_active_user), db:Session=Depends(get_db)):
     if db.query(User).filter(User.email==user.email).first():
         raise HTTPException(status_code=404, detail="User already exists!")
-    #create a new user
-    new_user = User(**user.dict())
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
+    
+    hashed_password = get_pwd_hash(user.password)
+    db_user = User(
+        name = user.name,
+        email = user.email,
+        role = user.role,
+        hashed_pwd = hashed_password
+    )
 
-    return new_user
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+
+    return db_user
 
 #Update User
 @app.put("/users/{user_id}", response_model=UserResponse)
-def update_user(user_id:int, user:UserCreate, db:Session=Depends(get_db)):
+def update_user(user_id:int, user:UserCreate,current_user:User=Depends(get_current_active_user) ,db:Session=Depends(get_db)):
     db_user = db.query(User).filter(User.id==user_id).first()
     if not db_user:
         raise HTTPException(status_code=404, detail="User not found!")
     
-    for field, value in user.dict().items():
-        setattr(db_user, field, value)
+    db_user.name = user.name
+    db_user.email = user.email
+    db_user.role =  user.role
+
 
     db.commit()
     db.refresh(db_user)
@@ -90,10 +254,14 @@ def update_user(user_id:int, user:UserCreate, db:Session=Depends(get_db)):
 
 #Delete User
 @app.delete("/users/{user_id}")
-def delete_user(user_id:int, db:Session=Depends(get_db)):
+def delete_user(user_id:int, current_user:User=Depends(get_current_active_user) ,db:Session=Depends(get_db)):
     db_user = db.query(User).filter(User.id==user_id).first()
     if not db_user:
         raise HTTPException(status_code=404, detail="User not found!")
+    
+    if db_user.id == current_user.id:
+        raise HTTPException(status_code=404, detail="You can't delete yourself")
+
     db.delete(db_user)
     db.commit()
 
@@ -101,6 +269,6 @@ def delete_user(user_id:int, db:Session=Depends(get_db)):
 
 #Get all Users
 @app.get("/users", response_model=List[UserResponse])
-def get_all_users(db:Session=Depends(get_db)):
+def get_all_users(current_user:User=Depends(get_current_active_user),db:Session=Depends(get_db)):
     return db.query(User).all()
